refactor(staging): unify factory functions, struct at top of file
Merge newS3StagingStoreFromConfig + newS3StagingStore into one
newS3StagingStore. Same for GCS. Each staging_*.go file now has:
struct → constructor → methods. staging.go is just the router.
